Created a mixin for report template ID, reorganized the pojos and created the delete and get commands.
