--- fcli-ssc/src/main/java/com/fortify/cli/ssc/common/ApplicationVersion.java
@@ -1,11 +1,0 @@
-package com.fortify.cli.ssc.common;
-
-import lombok.Getter;
-import lombok.Setter;
-
-public class ApplicationVersion {
-    @Getter @Setter String ApplicationId;
-    @Getter @Setter String ApplicationName;
-    @Getter @Setter String ApplicationVersionId;
-    @Getter @Setter String ApplicationVersionName;
-}
