--- fcli-ssc/src/main/java/com/fortify/cli/ssc/common/pojos/reportTemplateDef/newReportTemplate/ReportParameterOption.java
@@ -1,9 +1,0 @@
-package com.fortify.cli.ssc.common.pojos.reportTemplateDef.newReportTemplate;
-
-public class ReportParameterOption {
-    public boolean defaultValue;
-    public String description;
-    public String displayValue;
-    public int index;
-    public String reportValue;
-}
