--- fcli-ssc/src/main/java/com/fortify/cli/ssc/common/pojos/reportTemplateDef/newReportTemplate/enums/ReportParameterType.java
@@ -1,10 +1,0 @@
-package com.fortify.cli.ssc.common.pojos.reportTemplateDef.newReportTemplate.enums;
-
-public enum ReportParameterType {
-    BOOLEAN,
-    MULTI_PROJECT,
-    PROJECT_ATTRIBUTE,
-    SINGLE_PROJECT,
-    SINGLE_SELECT_DEFAULT,
-    STRING
-}
