--- fcli-ssc/src/main/java/com/fortify/cli/ssc/common/pojos/reportTemplateDef/newReportTemplate/enums/ReportRenderingEngineType.java
@@ -1,5 +1,0 @@
-package com.fortify.cli.ssc.common.pojos.reportTemplateDef.newReportTemplate.enums;
-
-public enum ReportRenderingEngineType {
-    BIRT
-}
