--- fcli-ssc/src/main/java/com/fortify/cli/ssc/common/pojos/reportTemplateDef/newReportTemplate/enums/ReportType.java
@@ -1,6 +1,0 @@
-package com.fortify.cli.ssc.common.pojos.reportTemplateDef.newReportTemplate.enums;
-
-public enum ReportType {
-    ISSUE,
-    PROJECT
-}
