+++ fcli-ssc/src/main/java/com/fortify/cli/ssc/picocli/command/appversion_artifact/SSCApplicationVersionArtifactDownloadCommand.java
@@ -26,8 +26,7 @@ package com.fortify.cli.ssc.picocli.command.appversion_artifact;
 
 import com.fortify.cli.common.picocli.mixin.output.IOutputConfigSupplier;
 import com.fortify.cli.common.picocli.mixin.output.OutputConfig;
-import com.fortify.cli.common.picocli.mixin.output.OutputMixin;
-import com.fortify.cli.ssc.common.ApplicationVersion;
+import com.fortify.cli.ssc.common.pojos.application.version.ApplicationVersion;
 import com.fortify.cli.ssc.common.SSCUrls;
 import com.fortify.cli.ssc.picocli.command.AbstractSSCUnirestRunnerCommand;
 import com.fortify.cli.ssc.picocli.mixin.application.version.SSCApplicationVersionIdMixin;
